+++ app/src/main/java/woowacourse/shopping/presentation/view/productdetail/ProductDetailActivity.kt
@@ -20,6 +20,7 @@ import woowacourse.shopping.presentation.model.CartModel
 import woowacourse.shopping.presentation.model.RecentProductModel
 import woowacourse.shopping.presentation.view.orderlist.OrderListActivity
 import woowacourse.shopping.presentation.view.productlist.ProductListActivity.Companion.KEY_SERVER_SERVER
+import woowacourse.shopping.presentation.view.util.createRetrofit
 import woowacourse.shopping.presentation.view.util.getParcelableCompat
 import woowacourse.shopping.presentation.view.util.getSerializableCompat
 import woowacourse.shopping.presentation.view.util.showToast
@@ -71,7 +72,8 @@ class ProductDetailActivity : AppCompatActivity(), ProductDetailContract.View {
     }
 
     private fun setPresenter() {
-        val productRemoteDataSource = ProductRemoteDataSourceImpl(server)
+        val retrofit = createRetrofit(server)
+        val productRemoteDataSource = ProductRemoteDataSourceImpl(retrofit)
         val cartLocalDataSource = CartLocalDataSourceImpl(this, server)
         val cartRemoteDataSource = CartRemoteDataSourceImpl(server)
         presenter = ProductDetailPresenter(
