+++ app/src/main/java/woowacourse/shopping/presentation/view/productlist/ProductListActivity.kt
@@ -32,6 +32,7 @@ import woowacourse.shopping.presentation.view.productlist.adpater.ProductListAda
 import woowacourse.shopping.presentation.view.productlist.adpater.RecentProductListAdapter
 import woowacourse.shopping.presentation.view.productlist.adpater.RecentProductWrapperAdapter
 import woowacourse.shopping.presentation.view.productlist.adpater.ViewType
+import woowacourse.shopping.presentation.view.util.createRetrofit
 import woowacourse.shopping.presentation.view.util.getSerializableCompat
 import woowacourse.shopping.presentation.view.util.showToast
 
@@ -129,7 +130,8 @@ class ProductListActivity : AppCompatActivity(), ProductContract.View {
     }
 
     private fun setPresenter() {
-        val productRemoteDataSource = ProductRemoteDataSourceImpl(server)
+        val retrofit = createRetrofit(server)
+        val productRemoteDataSource = ProductRemoteDataSourceImpl(retrofit)
         val cartLocalDataSource = CartLocalDataSourceImpl(this, server)
         val cartRemoteDataSource = CartRemoteDataSourceImpl(server)
         val recentProductLocalDataSource = RecentProductLocalDataSourceImpl(this, server)
